Move legacy-API form editors under Editors/legacyForm/
