+++ src/main/cliapp/src/components/ConditionRelationsEditDialog.jsx
@@ -8,8 +8,8 @@ import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { ValidationService } from '../service/ValidationService';
 import { DeleteAction } from './Actions/DeletionAction';
-import { InternalEditor } from './Editors/InternalEditor';
-import { ControlledVocabularyEditor } from './Editors/ControlledVocabularyEditor';
+import { InternalEditor } from './Editors/legacyForm/InternalEditor';
+import { ControlledVocabularyEditor } from './Editors/legacyForm/ControlledVocabularyEditor';
 import { DialogErrorMessageComponent } from './Error/DialogErrorMessageComponent';
 import { ExConAutocompleteTemplate } from './Editors/autocomplete/base/templates/ExConAutocompleteTemplate';
 import { SearchService } from '../service/SearchService';
