+++ src/main/cliapp/src/components/Editors/legacyForm/MutationTypesEditor.jsx
@@ -1,9 +1,9 @@
-import { AutocompleteMultiEditor } from './autocomplete/base/AutocompleteMultiEditor';
-import { SearchService } from '../../service/SearchService';
-import { autocompleteSearch, buildAutocompleteFilter } from '../../utils/utils';
-import { SubjectAutocompleteTemplate } from './autocomplete/base/templates/SubjectAutocompleteTemplate';
-import { DialogErrorMessageComponent } from '../Error/DialogErrorMessageComponent';
-import { Endpoints } from '../../constants/Endpoints';
+import { AutocompleteMultiEditor } from '../autocomplete/base/AutocompleteMultiEditor';
+import { SearchService } from '../../../service/SearchService';
+import { autocompleteSearch, buildAutocompleteFilter } from '../../../utils/utils';
+import { SubjectAutocompleteTemplate } from '../autocomplete/base/templates/SubjectAutocompleteTemplate';
+import { DialogErrorMessageComponent } from '../../Error/DialogErrorMessageComponent';
+import { Endpoints } from '../../../constants/Endpoints';
 
 const mutationTypeSearch = (event, setFiltered, setInputValue) => {
 	const searchService = new SearchService();
