+++ src/main/cliapp/src/components/Editors/legacyForm/ReferencesEditor.jsx
@@ -1,9 +1,9 @@
-import { AutocompleteMultiEditor } from './autocomplete/base/AutocompleteMultiEditor';
-import { SearchService } from '../../service/SearchService';
-import { autocompleteSearch, buildAutocompleteFilter } from '../../utils/utils';
-import { DialogErrorMessageComponent } from '../Error/DialogErrorMessageComponent';
-import { LiteratureAutocompleteTemplate } from './autocomplete/base/templates/LiteratureAutocompleteTemplate';
-import { Endpoints } from '../../constants/Endpoints';
+import { AutocompleteMultiEditor } from '../autocomplete/base/AutocompleteMultiEditor';
+import { SearchService } from '../../../service/SearchService';
+import { autocompleteSearch, buildAutocompleteFilter } from '../../../utils/utils';
+import { DialogErrorMessageComponent } from '../../Error/DialogErrorMessageComponent';
+import { LiteratureAutocompleteTemplate } from '../autocomplete/base/templates/LiteratureAutocompleteTemplate';
+import { Endpoints } from '../../../constants/Endpoints';
 
 const referenceSearch = (event, setFiltered, setInputValue) => {
 	const searchService = new SearchService();
