+++ src/main/cliapp/src/components/RelatedNotesEditDialog.jsx
@@ -8,9 +8,9 @@ import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { ValidationService } from '../service/ValidationService';
 import { DeleteAction } from './Actions/DeletionAction';
-import { InternalEditor } from './Editors/InternalEditor';
-import { ReferencesEditor } from './Editors/ReferencesEditor';
-import { VocabularyTermSetEditor } from './Editors/VocabularyTermSetEditor';
+import { InternalEditor } from './Editors/legacyForm/InternalEditor';
+import { ReferencesEditor } from './Editors/legacyForm/ReferencesEditor';
+import { VocabularyTermSetEditor } from './Editors/legacyForm/VocabularyTermSetEditor';
 import { TableInputTextAreaEditor } from './Editors/text/TableInputTextAreaEditor';
 import { Endpoints } from '../constants/Endpoints';
 import { multipleAutocompleteOnChange } from '../utils/utils';
