+++ src/main/cliapp/src/containers/allelesPage/fullName/FullNameFormTable.jsx
@@ -4,10 +4,10 @@ import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
 import { TableInputTextEditor } from '../../../components/Editors/text/TableInputTextEditor';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { ControlledVocabularyEditor } from '../../../components/Editors/ControlledVocabularyEditor';
-import { VocabularyTermSetEditor } from '../../../components/Editors/VocabularyTermSetEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { ControlledVocabularyEditor } from '../../../components/Editors/legacyForm/ControlledVocabularyEditor';
+import { VocabularyTermSetEditor } from '../../../components/Editors/legacyForm/VocabularyTermSetEditor';
 
 export const FullNameFormTable = ({
 	name,
