+++ src/main/cliapp/src/containers/allelesPage/functionalImpacts/FunctionalImpactsEditDialog.jsx
@@ -8,10 +8,10 @@ import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { ValidationService } from '../../../service/ValidationService';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { FunctionalImpactsEditor } from '../../../components/Editors/FunctionalImpactsEditor';
-import { PhenotypeTermEditor } from '../../../components/Editors/PhenotypeTermEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { FunctionalImpactsEditor } from '../../../components/Editors/legacyForm/FunctionalImpactsEditor';
+import { PhenotypeTermEditor } from '../../../components/Editors/legacyForm/PhenotypeTermEditor';
 import { TableInputTextAreaEditor } from '../../../components/Editors/text/TableInputTextAreaEditor';
 import { Endpoints } from '../../../constants/Endpoints';
 import { defaultAutocompleteOnChange, multipleAutocompleteOnChange } from '../../../utils/utils';
