+++ src/main/cliapp/src/containers/allelesPage/functionalImpacts/FunctionalImpactsFormTable.jsx
@@ -3,10 +3,10 @@ import { Column } from 'primereact/column';
 import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { FunctionalImpactsEditor } from '../../../components/Editors/FunctionalImpactsEditor';
-import { PhenotypeTermEditor } from '../../../components/Editors/PhenotypeTermEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { FunctionalImpactsEditor } from '../../../components/Editors/legacyForm/FunctionalImpactsEditor';
+import { PhenotypeTermEditor } from '../../../components/Editors/legacyForm/PhenotypeTermEditor';
 import { TableInputTextEditor } from '../../../components/Editors/text/TableInputTextEditor';
 
 export const FunctionalImpactsFormTable = ({
