+++ src/main/cliapp/src/containers/allelesPage/mutationTypes/MutationTypesFormTable.jsx
@@ -3,9 +3,9 @@ import { Column } from 'primereact/column';
 import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { MutationTypesEditor } from '../../../components/Editors/MutationTypesEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { MutationTypesEditor } from '../../../components/Editors/legacyForm/MutationTypesEditor';
 
 export const MutationTypesFormTable = ({
 	mutationTypes,
