+++ src/main/cliapp/src/containers/allelesPage/nomenclatureEvents/NomenclatureEventsFormTable.jsx
@@ -3,10 +3,10 @@ import { Column } from 'primereact/column';
 import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { ObsoleteEditor } from '../../../components/Editors/ObsoleteEditor';
-import { ControlledVocabularyEditor } from '../../../components/Editors/ControlledVocabularyEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { ObsoleteEditor } from '../../../components/Editors/legacyForm/ObsoleteEditor';
+import { ControlledVocabularyEditor } from '../../../components/Editors/legacyForm/ControlledVocabularyEditor';
 
 export const NomenclatureEventsFormTable = ({
 	nomenclatureEvents,
