+++ src/main/cliapp/src/containers/allelesPage/secondaryIds/SecondaryIdsFormTable.jsx
@@ -3,8 +3,8 @@ import { Column } from 'primereact/column';
 import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
 import { TableInputTextEditor } from '../../../components/Editors/text/TableInputTextEditor';
 
 export const SecondaryIdsFormTable = ({
