+++ src/main/cliapp/src/containers/nameSlotAnnotations/dialogs/FullNameEditDialog.jsx
@@ -10,9 +10,9 @@ import { DialogErrorMessageComponent } from '../../../components/Error/DialogErr
 import { ValidationService } from '../../../service/ValidationService';
 import { DeleteAction } from '../../../components/Actions/DeletionAction';
 import { TableInputTextEditor } from '../../../components/Editors/text/TableInputTextEditor';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { ControlledVocabularyEditor } from '../../../components/Editors/ControlledVocabularyEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { ControlledVocabularyEditor } from '../../../components/Editors/legacyForm/ControlledVocabularyEditor';
 import { useVocabularyTermSetService } from '../../../service/useVocabularyTermSetService';
 import { ControlledVocabularyDropdown } from '../../../components/Editors/dropdown/vocabulary/ControlledVocabularyDropdown';
