+++ src/main/cliapp/src/containers/nameSlotAnnotations/dialogs/SymbolEditDialog.jsx
@@ -9,9 +9,9 @@ import { Row } from 'primereact/row';
 import { DialogErrorMessageComponent } from '../../../components/Error/DialogErrorMessageComponent';
 import { ValidationService } from '../../../service/ValidationService';
 import { TableInputTextEditor } from '../../../components/Editors/text/TableInputTextEditor';
-import { InternalEditor } from '../../../components/Editors/InternalEditor';
-import { EvidenceEditor } from '../../../components/Editors/EvidenceEditor';
-import { ControlledVocabularyEditor } from '../../../components/Editors/ControlledVocabularyEditor';
+import { InternalEditor } from '../../../components/Editors/legacyForm/InternalEditor';
+import { EvidenceEditor } from '../../../components/Editors/legacyForm/EvidenceEditor';
+import { ControlledVocabularyEditor } from '../../../components/Editors/legacyForm/ControlledVocabularyEditor';
 import { useVocabularyTermSetService } from '../../../service/useVocabularyTermSetService';
 import { ControlledVocabularyDropdown } from '../../../components/Editors/dropdown/vocabulary/ControlledVocabularyDropdown';
